requirements.txt
aiofiles==23.2.1
aioredis==2.0.1
aiosqlite==0.17.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
Babel==2.14.0
bcrypt==4.1.2
black==23.12.1
cffi==1.16.0
click==8.1.7
cryptography==42.0.2
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
iso8601==1.1.0
isort==5.13.2
Jinja2==3.1.3
Mako==1.3.0
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.1.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic-extra-types==2.4.1
pydantic-settings==2.1.0
pydantic_core==2.14.6
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqladmin==0.16.0
SQLAlchemy==2.0.25
starlette==0.35.1
time-machine==2.13.0
tomli==2.0.1
tortoise-orm==0.20.0
typing_extensions==4.9.0
tzdata==2023.4
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
WTForms==3.1.2
